Add exec endpoint and move dns stuff to another package
